Add files via upload
This has the data of the residual capacity of an L.F.P type cell over a period of time. Use this data for reference purposes in case you want yo develop a model of your own. You can also verify the accuracy of my model in this .csv file.